+++ database/migrations/2020_06_16_071457_create_levels_table.php
@@ -18,7 +18,7 @@ class CreateLevelsTable extends Migration
             $table->string('name');
             $table->string('description');
             $table->integer('level');
-            $table->float('upgrade');
+            $table->double('upgrade');
             $table->timestamps();
         });
     }
