+++ database/migrations/2020_06_17_091349_create_cashbacks_table.php
@@ -17,7 +17,7 @@ class CreateCashbacksTable extends Migration
             $table->increments('id');
             $table->integer('offer_id');
             $table->integer('above')->nullable();
-            $table->float('percent')->nullable();
+            $table->double('percent')->nullable();
             $table->timestamps();
         });
     }
